LayoutTests/platform/mac/accessibility/meter-element-expected.txt

This tests that the meter element is accessible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Meter1
AXRole: AXProgressIndicator
AXTitle: 
AXDescription: 
AXValueDescription: 6 blocks used (out of 8 total)


Meter2
AXRole: AXProgressIndicator
AXTitle: 
AXDescription: 
AXValueDescription: 75%


Meter3
AXRole: AXProgressIndicator
AXTitle: 
AXDescription: 
AXValueDescription: 


Meter4
AXRole: AXProgressIndicator
AXTitle: 
AXDescription: 
AXValueDescription: 12cm


Meter5
AXRole: AXProgressIndicator
AXTitle: 
AXDescription: 
AXValueDescription: 2cm


Meter6
AXRole: AXProgressIndicator
AXTitle: 
AXDescription: 
AXValueDescription: 12cm


Meter7
AXRole: AXProgressIndicator
AXTitle: 
AXDescription: 
AXValueDescription: 2cm


PASS successfullyParsed is true

TEST COMPLETE
